move participantListReducer to original reducer folder
