+++ src/reducers/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from "redux";
-import participantReducer from "../components/ParticipantList/participantListReducer";
+import participantReducer from "./participantListReducer";
 
 export default combineReducers({
   participantList: participantReducer
